components names fixed
